+++ synapse/lib/stormtypes.py
@@ -8019,6 +8019,14 @@ class User(Prim):
                       {'name': 'default', 'type': 'boolean', 'desc': 'The default value.', 'default': False, },
                   ),
                   'returns': {'type': 'boolean', 'desc': 'True if the rule is allowed, False otherwise.', }}},
+        {'name': 'getAllowedReason', 'desc': 'Return an allowed status and reason for the given perm.',
+         'type': {'type': 'function', '_funcname': 'getAllowedReason',
+                  'args': (
+                      {'name': 'permname', 'type': 'str', 'desc': 'The permission string to check.', },
+                      {'name': 'gateiden', 'type': 'str', 'desc': 'The authgate iden.', 'default': None, },
+                      {'name': 'default', 'type': 'boolean', 'desc': 'The default value.', 'default': False, },
+                  ),
+                  'returns': {'type': 'list', 'desc': 'An (allowed, reason) tuple.', }}},
         {'name': 'grant', 'desc': 'Grant a Role to the User.',
          'type': {'type': 'function', '_funcname': '_methUserGrant',
                   'args': (
@@ -8209,6 +8217,7 @@ class User(Prim):
             'setEmail': self._methUserSetEmail,
             'setLocked': self._methUserSetLocked,
             'setPasswd': self._methUserSetPasswd,
+            'getAllowedReason': self.getAllowedReason,
         }
 
     async def _methUserPack(self):
@@ -8269,6 +8278,15 @@ class User(Prim):
         perm = tuple(permname.split('.'))
         user = await self.runt.snap.core.auth.reqUser(self.valu)
         return user.allowed(perm, gateiden=gateiden, default=default)
+
+    async def getAllowedReason(self, permname, gateiden=None, default=False):
+        permname = await tostr(permname)
+        gateiden = await tostr(gateiden)
+        default = await tobool(default)
+
+        perm = tuple(permname.split('.'))
+        user = await self.runt.snap.core.auth.reqUser(self.valu)
+        return user.getAllowedReason(perm, gateiden=gateiden, default=default)
 
     async def _methUserGrant(self, iden):
         self.runt.confirm(('auth', 'user', 'grant'))
@@ -9347,6 +9365,10 @@ async def toiter(valu, noneok=False):
 async def torepr(valu, usestr=False):
     if hasattr(valu, 'stormrepr') and callable(valu.stormrepr):
         return await valu.stormrepr()
+
+    if isinstance(valu, bool):
+        return str(valu).lower()
+
     if usestr:
         return str(valu)
     return repr(valu)